refactor(Pieces): Pieces 분리
