+++ src/main/java/domain/Game.java
@@ -8,45 +8,10 @@ import domain.position.Row;
 import java.util.*;
 
 public class Game {
-    private final List<Piece> pieces = new ArrayList<>();
+    private final Pieces pieces;
 
     public Game() {
-        setUpGeneral();
-        setUpPawn();
-    }
-
-    private void setUpGeneral() {
-        setUpGeneralByColor(Color.BLACK, Row.EIGHT);
-        setUpGeneralByColor(Color.WHITE, Row.ONE);
-    }
-
-    private void setUpGeneralByColor(Color color, Row row) {
-        pieces.add(new Rook(color, Position.of(Column.A, row)));
-        pieces.add(new Knight(color, Position.of(Column.B, row)));
-        pieces.add(new Bishop(color, Position.of(Column.C, row)));
-        pieces.add(new Queen(color, Position.of(Column.D, row)));
-        pieces.add(new King(color, Position.of(Column.E, row)));
-        pieces.add(new Bishop(color, Position.of(Column.F, row)));
-        pieces.add(new Knight(color, Position.of(Column.G, row)));
-        pieces.add(new Rook(color, Position.of(Column.H, row)));
-    }
-
-    public void setUpPawn() {
-        setUpRow(Row.SEVEN, Color.BLACK);
-        setUpRow(Row.TWO, Color.WHITE);
-    }
-
-    private void setUpRow(Row row, Color color) {
-        for (Column column : Column.values()) {
-            pieces.add(new Pawn(color, Position.of(column, row)));
-        }
-    }
-
-    private Piece hasPosition(Row row, Column column) {
-        return pieces.stream()
-                     .filter(piece -> piece.hasPosition(Position.of(column, row)))
-                     .findFirst()
-                     .orElse(new Empty());
+        pieces = new Pieces();
     }
 
     public void display() {
@@ -54,7 +19,7 @@ public class Game {
         Collections.reverse(rows);
         for (Row row : rows) {
             for (Column column : Column.values()) {
-                System.out.print(hasPosition(row, column).display());
+                System.out.print(pieces.getPieceOf(row, column).display());
             }
             System.out.println();
         }
