+++ src/test/java/domain/GameTest.java
@@ -2,7 +2,6 @@ package domain;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
 
 class GameTest {
     @Test
